% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2pedgene.R
\name{DOpedgene}
\alias{DOpedgene}
\title{pedgene call back function}
\usage{
DOpedgene(geno_arg, markers_arg, range_arg, envir = ENV)
}
\arguments{
\item{geno_arg}{A matrix with one row per \emph{fam} pedigree member and one column for each marker that is selected.  Each datum are the two characters indicating the genotype for the marker/member.}

\item{markers_arg}{a data.frame with the following 5 variables:
\describe{
\item{locus_link}{is the ordinal ranking of this marker among all loci}
\item{locus_link_fill}{is the position of corresponding genotype data in the
\emph{unified_genotype_table}}
\item{MarkerName}{is the text name of the marker}
\item{chromosome}{is the integer chromosome number}
\item{position}{is the integer base pair position of marker}
 }}

\item{range_arg}{one row of a ranges_arg.  The latter is a data.frame of at least three
integer columns.  The columns indicate a range:
a chromosome number, a start base pair value, and an end base pair value.}

\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
...
}
\examples{
\dontrun{
   applyFnToRanges(DOpedgene, ENV$refRanges[gs, ], ENV$refIndices, ENV)
}
}
