% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2pedgene.R
\docType{package}
\name{Mega2pedgene-package}
\alias{Mega2pedgene-package}
\title{Mega2pedgene package}
\description{
This package runs the \bold{pedgene} program via the \bold{Mega2} applyFnToRanges Framework.
 The results are stored in a data frame and can then be written to disk if desired.
}
\references{
https://cran.r-project.org/web/packages/pedgene/index.html
}
\author{
Robert V Baron, rvb5@pitt.edu
}
