% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2pedgene.R
\name{DOpedgene}
\alias{DOpedgene}
\title{pedgene call back function}
\usage{
DOpedgene(geno_arg, markers_arg, range_arg, envir = ENV)
}
\arguments{
\item{geno_arg}{A character matrix with one row per \emph{fam} pedigree member and one column for each marker in
markers_arg.  Each cell contains the two characters indicating the nucleotides for the marker.}

\item{markers_arg}{a data.frame with the following 5 variables:
\describe{
\item{locus_link}{is the ordinal ranking of this marker among all loci}
\item{locus_link_fill}{is the position of corresponding genotype data in the
\emph{unified_genotype_table}}
\item{MarkerName}{is the text name of the marker}
\item{chromosome}{is the integer chromosome number}
\item{position}{is the integer base pair position of marker}
 }}

\item{range_arg}{one row of a ranges_arg.  The latter is a data frame of at least three
integer columns.  The columns indicate a range:
a chromosome number, a start base pair value, and an end base pair value.}

\item{envir}{"R environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
First, ignore call backs that have less than two markers.  Second, convert the genotypes
 patterns of 1/1, 1/2 (and 2/1) and 2/2 in the genotype matrix
 to the numbers 0, 1, 2 for each marker. (Reverse, the order iff allele "1" has the
 minor allele frequency.)  Finally, prepend the pedigree and person columns of the family data
 to processed genotype matrix.  Finally, invoke \code{pedgene} with the family data and converted
 genotype matrix for several different weights.  Save the kernel and burden, value and p-value for each
 measurement in \emph{envir$pedgene_results}.
}
\note{
This function accumulates output in the data frame, \emph{envir$pedgene_results}.  It will
 print out the lines as they are generated if \emph{envir$verbose} is TRUE.  It does not write anything
 to a file.  You must save the data frame or the "observations" you need by yourself.
}
\examples{
\dontrun{
   applyFnToRanges(DOpedgene, ENV$refRanges[gs, ], ENV$refIndices, ENV)
}
}
