% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2pedgene.R
\name{init_pedgene}
\alias{init_pedgene}
\title{load Mega2 SQLite database into R and perform initialization for pedgene usage}
\usage{
init_pedgene(db = NULL, filename = NULL, verbose = FALSE)
}
\arguments{
\item{db}{specifies a \bold{Mega2} SQLite database containing study data.}

\item{filename}{in which to store results data frame.  By default "pedgene.txt" is used.}

\item{verbose}{TRUE indicates that diagnostic printouts should be enabled.
This value is saved in the returned environment.}
}
\value{
"environment" containing data frames from an SQLite database and some computed values.
}
\description{
This populates the \bold{R} data frames from the specified \b{Mega2 R} database.  It then
 prunes the samples to only include members that have a definite case or control
 status.  Undefined samples are ignored; this is necessary for \emph{pedgene}.
}
\note{
This also calculates schaidPed and pedPer that are used later in the \emph{Dopedgene} calculation.

 This also initializes the dataframe "envir$pedgene_results" to zero rows.
}
\examples{
\dontrun{
init_pedgene("ped3.db", verbose = TRUE)
}
}
