% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2pedgene.R
\name{run_pedgene}
\alias{run_pedgene}
\title{execute the pedgene function on a subset of the gene transcript ranges}
\usage{
run_pedgene(gs = 1:100, envir = ENV)
}
\arguments{
\item{gs}{a subsequence of the gene transcript ranges to calculate the \emph{Dopedgene} function
on.}

\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
execute the pedgene function on a subset of the gene transcript ranges
}
\note{
This code starts by deleting the output file ("pedgene.txt" by default).  After Dopedgene 
 is applied to all the chosen transcripts.  The data frame "envir$pedgene_results" is written
 to the output file.
}
\examples{
\dontrun{
run()

run(1:10)
}
}
