% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2pedgene.R
\name{run_pedgene}
\alias{run_pedgene}
\title{execute the CRAN pedgene function on a subset of the default gene transcript ranges}
\usage{
run_pedgene(gs = 1:100, envir = ENV)
}
\arguments{
\item{gs}{a subrange of the default transcript ranges over which to calculate the \emph{Dopedgene} function.}

\item{envir}{"R environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
Execute the pedgene function on the first \emph{gs} default gene transcript ranges (gs = 1:100).
 Update the \emph{envir$pedgene_results} data frame with the results.
}
\note{
This code starts by deleting the output file set in \code{init_pedgene} ("pedgene.txt" by default).  Then \code{Dopedgene}
 is applied to all the appropriate ranges.  Finally, the data frame of results, \emph{envir$pedgene_results}, is written
 to the output file.
}
\examples{
\dontrun{
run_pedgene()

run_pedgene(1:10)
}
}
