% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2pedgene.R
\name{run_pedgene}
\alias{run_pedgene}
\title{execute the pedgene function on a subset of the gene transcript ranges}
\usage{
run_pedgene(gs = 1:100)
}
\arguments{
\item{gs}{a subsequence of the gene transcript ranges to calculate the \emph{Dopedgene} function
on.}
}
\value{
None
}
\description{
execute the pedgene function on a subset of the gene transcript ranges
}
\examples{
\dontrun{
run()

run(1:10)
}
}
